Refactor: Host 팀 화면 레이아웃을 공통 셸 구조로 정렬)
Main Shell에서 팀 화면을 렌더링할 때 공통 사이드바(primary)와 팀 전용 보조 패널(secondary)의 책임을 분리하도록 레이아웃 구성을 정리했습니다.
팀 화면에서만 발생하던 사이드바 디자인 불일치 문제를 줄이기 위해 host 측 렌더 경로와 페이지 레이아웃 진입점을 통일하고, pages 라우터 환경에서의 문서/앱 래핑 구조를 보강했습니다.
이 변경은 팀 전용 UI를 secondary 패널 슬롯으로 수용하는 기반 작업이며, 기존 공통 셸 디자인 톤과 동작을 유지하면서 팀 화면만 구조적으로 분리되도록 맞췄습니다.
