+++ apps/web/src/lib/host-document-fonts.ts
@@ -1,6 +1,6 @@
 import { Geist, Geist_Mono } from "next/font/google";
 
-/** identity/team-web 과 동일한 Geist 변수 — Pages Router는 `layout`이 없어 `_app`에서 `documentElement`에 부착한다(Task37-20). */
+/** identity/team-web 과 동일한 Geist 변수 — Pages 호스트는 `pages/_document.tsx`의 `<Html>`에 부착한다. */
 export const hostGeistSans = Geist({
   variable: "--font-geist-sans",
   subsets: ["latin"],
